# Support wildcard subscriptions + getter for known types (optimized)