--- a/go.mod
+++ b/go.mod
@@ -3,6 +3,11 @@ module github.com/libp2p/go-eventbus
 go 1.12
 
 require (
-	github.com/libp2p/go-libp2p-core v0.5.0
+	github.com/ipfs/go-cid v0.0.5 // indirect
+	github.com/libp2p/go-buffer-pool v0.0.2 // indirect
+	github.com/libp2p/go-libp2p-core v0.3.2-0.20200415071619-8fb76f42df50
 	github.com/libp2p/go-libp2p-testing v0.1.1
+	github.com/multiformats/go-multiaddr v0.2.1 // indirect
+	github.com/stretchr/testify v1.4.0
+	go.opencensus.io v0.22.3 // indirect
 )