--- a/go.mod
+++ b/go.mod
@@ -5,7 +5,7 @@ go 1.12
 require (
 	github.com/ipfs/go-cid v0.0.5 // indirect
 	github.com/libp2p/go-buffer-pool v0.0.2 // indirect
-	github.com/libp2p/go-libp2p-core v0.3.2-0.20200416170247-0a7a9856715c
+	github.com/libp2p/go-libp2p-core v0.5.2-0.20200417060929-6957bf8a421d
 	github.com/libp2p/go-libp2p-testing v0.1.1
 	github.com/multiformats/go-multiaddr v0.2.1 // indirect
 	github.com/stretchr/testify v1.4.0