--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,7 @@ module github.com/libp2p/go-eventbus
 go 1.12
 
 require (
-	github.com/libp2p/go-libp2p-core v0.5.2-0.20200417060929-6957bf8a421d
+	github.com/libp2p/go-libp2p-core v0.5.2-0.20200520121154-a1698eb742b1
 	github.com/libp2p/go-libp2p-testing v0.1.1
 	github.com/stretchr/testify v1.4.0
 )
-
-replace github.com/libp2p/go-libp2p-core => ../go-libp2p-core